+++ api-builder-plugin-fn-aws-athena/test/config/aws-athena.testconfig.js
@@ -5,8 +5,8 @@ module.exports = {
 			'aws': {
 				credentials: {
 					region: "us-east-1",
-					accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-					secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY
+					accessKeyId: 'MOCK',
+					secretAccessKey: 'MOCK'
 				}
 			}
 		}