+++ api-builder-plugin-fn-aws-athena/test/test.js
@@ -1,5 +1,6 @@
 const { expect } = require('chai');
 const { MockRuntime } = require('@axway/api-builder-sdk');
+var simple = require('simple-mock');
 
 const getPlugin = require('../src');
 const actions = require('../src/actions');
@@ -24,35 +25,8 @@ describe('Basic configuration tests for AWS Athena connector', () => {
 			expect(flownode.icon).to.be.a('string');
 		});
 
-		// It is vital to ensure that the generated node flow-nodes are valid
-		// for use in API Builder. Your unit tests should always include this
-		// validation to avoid potential issues when API Builder loads your
-		// node.
 		it('should define valid flow-nodes', () => {
 			expect(runtime.validate()).to.not.throw;
-		});
-	});
-});
-
-describe('Incomplete PluginConfig tests', () => {
-	let runtime;
-	before(async () => runtime = new MockRuntime(await getPlugin(invalidPluginConfig)));
-
-	describe('Check config file', () => {
-		it('Validate error handling without config file configured', async () => {
-			const flowNode = runtime.getFlowNode('athena');
-			const result = await flowNode.query({
-				table: null
-			});
-
-			expect(result.callCount).to.equal(1);
-			expect(result.output).to.equal('error');
-			expect(result.args[0]).to.equal(null);
-			expect(result.args[1]).to.be.instanceOf(Error)
-				.and.to.have.property('message', 'Your AWS-Athena configuration file is incomplete. Please check conf/aws-athena.default.js');
-			expect(result.context).to.be.an('Object');
-			expect(result.context.error).instanceOf(Error)
-				.and.to.have.property('message', 'Your AWS-Athena configuration file is incomplete. Please check conf/aws-athena.default.js');
 		});
 	});
 });
@@ -113,6 +87,17 @@ describe('Valid PluginConfig tests', () => {
 		it('Make a simple query without a given limit - Must be limited to default 5', async () => {
 			const flowNode = runtime.getFlowNode('athena');
 
+			simple.mock(runtime.plugin.flownodes.athena.athenaClient, 'query').callFn(function (params) {
+				expect(params.sql).to.be.an('String');
+				expect(params.sql).to.equals('SELECT elb_name, request_ip FROM elb_logs LIMIT 5');
+				expect(params.db).to.be.an('String');
+				expect(params.db).to.equals('sampledb');
+				return Promise.resolve( { Items: [
+					{elb_name: "elb_demo_003", request_ip: "251.165.102.100"}, 
+					{elb_name: "elb_demo_007", request_ip: "250.120.176.53"}, 
+				]});
+			});
+
 			const result = await flowNode.query({ table: 'elb_logs', db: 'sampledb', fields: 'elb_name, request_ip'});
 
 			expect(result.callCount).to.equal(1);
@@ -121,11 +106,22 @@ describe('Valid PluginConfig tests', () => {
 			expect(result.args[1]).to.have.keys(['Items']);
 			expect(result.args[1].Items[0]).to.have.any.keys(['elb_name', 'request_ip']);
 			expect(result.args[1].Items[0]).to.not.have.any.keys(['request_port']);
-			expect(result.args[1].Items).to.have.lengthOf(5);
-		}).timeout(10000);
+			expect(result.args[1].Items).to.have.lengthOf(2);
+		});
 
 		it('Make a simple query with limit 1', async () => {
+			debugger;
 			const flowNode = runtime.getFlowNode('athena');
+
+			simple.mock(runtime.plugin.flownodes.athena.athenaClient, 'query').callFn(function (params) {
+				expect(params.sql).to.be.an('String');
+				expect(params.sql).to.equals('SELECT elb_name, request_port FROM elb_logs LIMIT 1');
+				expect(params.db).to.be.an('String');
+				expect(params.db).to.equals('sampledb');
+				return Promise.resolve( { Items: [
+					{elb_name: "elb_demo_003", request_port: "24588"}
+				]});
+			});
 
 			const result = await flowNode.query({ table: 'elb_logs', db: 'sampledb', fields: 'elb_name, request_port', limit: '1' });